chore:rename servefile to serveIndex
